[be/refactor] 가격 관련 데이터 자료형 변경 (Integer -> Long)

See merge request s08-bigdata-dist-sub2/S08P22A202!215

+++ backend/auction/src/main/java/com/dokidoki/auction/dto/db/MostSaleProductInterface.java
@@ -2,5 +2,5 @@ package com.dokidoki.auction.dto.db;
 
 public interface MostSaleProductInterface {
     String getName();
-    Integer getSale_cnt();
+    Long getSale_cnt();
 }
